+++ packages/repo-tools/src/test-utils/index.ts
@@ -2,6 +2,7 @@ export { delay } from './delay.ts';
 export { makeErrorMatcherFactory } from './errors.ts';
 export { makePromiseKitMock } from './promise-kit.ts';
 export { fetchMock } from './env/fetch-mock.ts';
+export { makeTwoFacedFetchInput } from './fetch-input.ts';
 export * from './env/mock-kernel.ts';
 export { makeMockMessageTarget } from './postMessage.ts';
 export { makeAbortSignalMock } from './abort-signal.ts';